% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geno}
\alias{get_geno}
\title{Get genotypes from haplotypes}
\usage{
get_geno(hap, genotypes = c("NN", "NP", NA))
}
\arguments{
\item{hap}{haplotypes data with format: X/Y:n,m}

\item{genotypes}{genotypes to assign based on homozygosity,
c(Homozygous, Heterozygous, NULL); NULL haplotypes \link{./.:0}}
}
\value{
genotype
}
\description{
Get genotypes from haplotypes
}
\examples{
get_geno("./.:0")
get_geno("4/2:34,24")
get_geno("2/2:71")
}
