% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex_logo}
\alias{hex_logo}
\title{Create hexagonal logo for the package}
\usage{
hex_logo(
  subplot = system.file("images/slope.png", package = "RhAMPseq"),
  dpi = 600,
  h_color = "#000000",
  h_fill = "#363B74",
  output = system.file("images/logo.png", package = "RhAMPseq"),
  p_color = "#EEEEEE",
  url = "https://github.com/villegar/RhAMPseq",
  u_size = 1.35
)
}
\arguments{
\item{subplot}{image to use as the main logo}

\item{dpi}{plot resolution (dots-per-inch)}

\item{h_color}{color for hexagon border}

\item{h_fill}{color to fill hexagon}

\item{output}{output file (hexagonal logo)}

\item{p_color}{color for package name}

\item{url}{URL for package repository or website}

\item{u_size}{text size for URL}

\item{package}{title for logo (package name)}
}
\value{
hexagonal logo
}
\description{
Create hexagonal logo for the package
}
\examples{
hex_logo()
\dontrun{
hex_logo("inst/images/slope.png", output = "inst/images/logo.png")
}
}
