% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load rhAmpSeq raw data in different formats}
\usage{
load_data(fasta = NULL, hap_geno = NULL, count_mat = NULL)
}
\arguments{
\item{fasta}{FASTA data}

\item{hap_geno}{genotypes data}

\item{count_mat}{count matrix}
}
\value{
list of each loaded file
}
\description{
Load rhAmpSeq raw data in different formats
}
\examples{
\dontrun{
load_data(fasta = "data/HaplotypeAllele.fasta",
          hap_geno = "data/hap_genotype",
          count_mat = "data/readCountMatrixFile")
}
}
