% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_excel_col}
\alias{read_excel_col}
\title{Read columns from MS Excel file}
\usage{
read_excel_col(filename, columns)
}
\arguments{
\item{filename}{filename, absolute or relative paths are valid}

\item{columns}{vector with the columns names}
}
\value{
table with the requested columns
}
\description{
Read columns from MS Excel file
}
\examples{
\dontrun{
read_excel_col("data/Rhampseq_populations.xlsx", "A")
}
}
