% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_haplo}
\alias{split_haplo}
\title{Split haplotypes}
\usage{
split_haplo(hap)
}
\arguments{
\item{hap}{haplotype data with the format: \code{X/Y:n,m}.}
}
\value{
A tibble with three columns: \code{A}, \code{B}, and \code{len}.
}
\description{
Split haplotypes into three components: \code{A/B,len}.
}
\examples{
split_haplo("./.:0")
split_haplo("4/2:34,24")
split_haplo("2/2:71")
split_haplo("4/2:2,1")
split_haplo("2/2:4")
}
